+++ boards/m4a-24g/include/board.h
@@ -25,7 +25,6 @@
 #include "periph_conf.h"
 #include "periph_cpu.h"
 #include "periph/cpuid.h"
-#include "uniqueid.h"
 
 #ifdef __cplusplus
 extern "C" {
