+++ boards/m4a-wroom/include/board.h
@@ -24,7 +24,6 @@
 #include "periph_conf.h"
 #include "periph_cpu.h"
 #include "periph/cpuid.h"
-#include "uniqueid.h"
 
 /**
  * @name    Button pin definitions
