+++ firmware/sys/Makefile
@@ -3,4 +3,6 @@ ifneq (,$(filter storage,$(USEMODULE)))
   DIRS += storage
 endif
 
+DIRS += network
+
 include $(RIOTBASE)/Makefile.base
