+++ firmware/sys/network/Makefile
@@ -1,1 +1,5 @@
+DIRS += uniqueid
+DIRS += udp_client
+DIRS += udp_server
+
 include $(RIOTBASE)/Makefile.base
