+++ firmware/sys/network/Makefile.dep
@@ -1,2 +1,1 @@
-USEMODULE += random
-USEMODULE += xtimer
+USEMODULE += uniqueid
