--- firmware/sys/network/Makefile.include
@@ -1,2 +1,0 @@
-USEMODULE_INCLUDES_network := $(LAST_MAKEFILEDIR)/include
-USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_network)
