+++ firmware/sys/network/udp_client/Makefile
@@ -1,1 +1,2 @@
+SCR= udp_client.c
 include $(RIOTBASE)/Makefile.base
