+++ firmware/sys/network/udp_client/Makefile.include
@@ -1,0 +1,2 @@
+USEMODULE_INCLUDES_udp_client := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_udp_client)
