+++ firmware/sys/network/udp_server/Makefile
@@ -1,0 +1,2 @@
+SCR= udp_server.c
+include $(RIOTBASE)/Makefile.base
