+++ firmware/sys/network/udp_server/Makefile.include
@@ -1,0 +1,2 @@
+USEMODULE_INCLUDES_udp_server := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_udp_server)
