+++ firmware/sys/network/udp_server/udp_server.c
@@ -25,7 +25,7 @@
 #include "net/sock/udp.h"
 #include "net/ipv6/addr.h"
 #include "thread.h"
-#include "udp_m4a.h"
+#include "udp_server.h"
 
 #define SERVER_MSG_QUEUE_SIZE (8)
 #define SERVER_BUFFER_SIZE (100)
@@ -60,28 +60,6 @@ void *_udp_serverf(void *args) {
     }
 
     return NULL;
-}
-
-int udp_send(int *port, char *address, uint8_t *message, size_t *payload_len) {
-    int res;
-    sock_udp_ep_t remote = {.family = AF_INET6};
-
-    if (ipv6_addr_from_str((ipv6_addr_t *)&remote.addr, address) == NULL) {
-        puts("Error: unable to parse destination address");
-        return 1;
-    }
-    if (ipv6_addr_is_link_local((ipv6_addr_t *)&remote.addr)) {
-        /* choose first interface when address is link local */
-        gnrc_netif_t *netif = gnrc_netif_iter(NULL);
-        remote.netif = (uint16_t)netif->pid;
-    }
-    remote.port = *port;
-    if ((res = sock_udp_send(NULL, message, *payload_len, &remote)) < 0) {
-        puts("could not send");
-    } else {
-        printf("Success: send %u bytes  from %s \n", (unsigned)res, address);
-    }
-    return 0;
 }
 
 int udp_server(udpf_payload *argv) {
