+++ firmware/sys/network/uniqueid/Makefile.dep
@@ -1,0 +1,2 @@
+USEMODULE += random
+USEMODULE += xtimer
