+++ firmware/sys/network/uniqueid/Makefile.include
@@ -1,0 +1,2 @@
+USEMODULE_INCLUDES_uniqueid := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_uniqueid)
