--- firmware/sys/udp_m4a/Makefile.include
@@ -1,2 +1,0 @@
-USEMODULE_INCLUDES_UDP_M4A := $(LAST_MAKEFILEDIR)/include
-USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_UDP_M4A)
